[*] Implemented Bad Solution
This solutions changes the validation code to the Reservation class, and, in case of a Exception: the class returns a String containing information about the Exception.
Also, a validation still yet at the main class (Line 31). This validation can't be passed to the Reservation class, because the validation would have to be on the constructor, who doesn't return anything. Making this method not so good.