[+] Implemented good method.
This method uses exceptions with a custom Exception (not necessary to be a custom one, but it's cooler to use it) to treat possible errors. All the validation code is made on the Reservation class. Also, I added back the .gitignore file in the previous commit, which I accidentally removed in the first one. xD